+++ HISTORY.rst
@@ -7,6 +7,9 @@ History
 - Remove 3Di from user facing text, replace with Rana
 - Remove Lutra copyright statements
 - Replace icons for Rana-style icons
+- Fix bug with  "use selected features" (#469)
+- Make crest_level en crest_type optional for Weir and Orifice (#471)
+
 
 2.4.3 (2025-11-26)
 ------------------